refactor(tests): Split test_environment.py
